+++ lwpkt/src/include/lwpkt/lwpkt.h
@@ -4,7 +4,7 @@
  */
 
 /*
- * Copyright (c) 2020 Tilen MAJERLE
+ * Copyright (c) 2024 Tilen MAJERLE
  *
  * Permission is hereby granted, free of charge, to any person
  * obtaining a copy of this software and associated documentation
@@ -29,15 +29,15 @@
  * This file is part of LwPKT - Lightweight packet protocol library.
  *
  * Author:          Tilen MAJERLE <tilen@majerle.eu>
- * Version:         v1.2.0
+ * Version:         v1.3.0
  */
 #ifndef LWPKT_HDR_H
 #define LWPKT_HDR_H
 
-#include <string.h>
 #include <stdint.h>
-#include "lwrb/lwrb.h"
+#include <string.h>
 #include "lwpkt/lwpkt_opt.h"
+#include "lwrb/lwrb.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -53,35 +53,37 @@ extern "C" {
  * \brief           Packet state enumeration
  */
 typedef enum {
-    LWPKT_STATE_START = 0x00,                   /*!< Packet waits for start byte */
-    LWPKT_STATE_FROM,                           /*!< Packet waits for "packet from" byte */
-    LWPKT_STATE_TO,                             /*!< Packet waits for "packet to" byte */
-    LWPKT_STATE_CMD,                            /*!< Packet waits for "packet cmd" byte */
-    LWPKT_STATE_LEN,                            /*!< Packet waits for (multiple) data length bytes */
-    LWPKT_STATE_DATA,                           /*!< Packet waits for actual data bytes */
-    LWPKT_STATE_CRC,                            /*!< Packet waits for CRC data */
-    LWPKT_STATE_STOP,                           /*!< Packet waits for stop byte */
+    LWPKT_STATE_START = 0x00, /*!< Packet waits for start byte */
+    LWPKT_STATE_FROM,         /*!< Packet waits for "packet from" byte */
+    LWPKT_STATE_TO,           /*!< Packet waits for "packet to" byte */
+    LWPKT_STATE_CMD,          /*!< Packet waits for "packet cmd" byte */
+    LWPKT_STATE_FLAGS,        /*!< Packet waits for "packet flags" byte (custom user flags) */
+    LWPKT_STATE_LEN,          /*!< Packet waits for (multiple) data length bytes */
+    LWPKT_STATE_DATA,         /*!< Packet waits for actual data bytes */
+    LWPKT_STATE_CRC,          /*!< Packet waits for CRC data */
+    LWPKT_STATE_STOP,         /*!< Packet waits for stop byte */
+    LWPKT_STATE_END,          /*!< Last entry */
 } lwpkt_state_t;
 
 /**
  * \brief           Packet result enumeration
  */
 typedef enum {
-    lwpktOK = 0x00,                             /*!< Function returns successfully */
-    lwpktERR,                                   /*!< General error for function status */
-    lwpktINPROG,                                /*!< Receive is in progress */
-    lwpktVALID,                                 /*!< packet valid and ready to be read as CRC is valid and STOP received */
-    lwpktERRCRC,                                /*!< CRC integrity error for the packet. Will not wait STOP byte if received */
-    lwpktERRSTOP,                               /*!< Packet error with STOP byte, wrong character received for STOP */
-    lwpktWAITDATA,                              /*!< Packet state is in start mode, waiting start byte to start receiving */
-    lwpktERRMEM,                                /*!< No enough memory available for write */
+    lwpktOK = 0x00, /*!< Function returns successfully */
+    lwpktERR,       /*!< General error for function status */
+    lwpktINPROG,    /*!< Receive is in progress */
+    lwpktVALID,     /*!< packet valid and ready to be read as CRC is valid and STOP received */
+    lwpktERRCRC,    /*!< CRC integrity error for the packet. Will not wait STOP byte if received */
+    lwpktERRSTOP,   /*!< Packet error with STOP byte, wrong character received for STOP */
+    lwpktWAITDATA,  /*!< Packet state is in start mode, waiting start byte to start receiving */
+    lwpktERRMEM,    /*!< No enough memory available for write */
 } lwpktr_t;
 
 /**
  * \brief           CRC structure for packet
  */
 typedef struct {
-    uint8_t crc;                                /*!< Current CRC value */
+    uint8_t crc; /*!< Current CRC value */
 } lwpkt_crc_t;
 
 /* Forward declaration */
@@ -91,12 +93,24 @@ struct lwpkt;
  * \brief           List of event types
  */
 typedef enum {
-    LWPKT_EVT_PKT,                              /*!< Valid packet ready to read */
-    LWPKT_EVT_TIMEOUT,                          /*!< Timeout on packat, reset event */
-    LWPKT_EVT_READ,                             /*!< Packet read operation.
-                                                         Called when read operation happens from RX buffer */
-    LWPKT_EVT_WRITE,                            /*!< Packet write operation.
-                                                         Called when write operation happens to TX buffer  */
+    LWPKT_EVT_PKT,        /*!< Valid packet ready to read */
+    LWPKT_EVT_TIMEOUT,    /*!< Timeout on packat, reset event */
+    LWPKT_EVT_READ,       /*!< Packet read operation.
+                            Called when read operation happens from RX buffer */
+    LWPKT_EVT_WRITE,      /*!< Packet write operation.
+                            Called when write operation happens to TX buffer  */
+    LWPKT_EVT_PRE_WRITE,  /*!< Packet pre-write operation.
+                                Called before write operation could even start.
+                                It can be used to get exclusive mutex access to the resource */
+    LWPKT_EVT_POST_WRITE, /*!< Packet post-write operation.
+                                Called after write operation finished.
+                                It can be used to release exclusive mutex access from the resource */
+    LWPKT_EVT_PRE_READ,   /*!< Packet pre-read operation.
+                                Called before read operation could even start.
+                                It can be used to get exclusive mutex access to the resource */
+    LWPKT_EVT_POST_READ,  /*!< Packet post-read operation.
+                                Called after read operation finished.
+                                It can be used to release exclusive mutex access from the resource */
 } lwpkt_evt_type_t;
 
 /**
@@ -120,96 +134,117 @@ typedef uint8_t lwpkt_addr_t;
  */
 typedef struct lwpkt {
 #if LWPKT_CFG_USE_ADDR || __DOXYGEN__
-    lwpkt_addr_t addr;                          /*!< Current device address */
-#endif /* LWPKT_CFG_USE_ADDR || __DOXYGEN__ */
-    uint8_t data[LWPKT_CFG_MAX_DATA_LEN];       /*!< Memory to write received data */
-    LWRB_VOLATILE lwrb_t* tx_rb;                /*!< TX ringbuffer */
-    LWRB_VOLATILE lwrb_t* rx_rb;                /*!< RX ringbuffer */
-    uint32_t last_rx_time;                      /*!< Last RX time in units of milliseconds */
+    lwpkt_addr_t addr;                    /*!< Current device address */
+#endif                                    /* LWPKT_CFG_USE_ADDR || __DOXYGEN__ */
+    uint8_t data[LWPKT_CFG_MAX_DATA_LEN]; /*!< Memory to write received data */
+    lwrb_t* tx_rb;                        /*!< TX ringbuffer */
+    lwrb_t* rx_rb;                        /*!< RX ringbuffer */
+    uint32_t last_rx_time;                /*!< Last RX time in units of milliseconds */
 #if LWPKT_CFG_USE_EVT || __DOXYGEN__
-    lwpkt_evt_fn evt_fn;                        /*!< Global event function for read and write operation */
-#endif /* LWPKT_CFG_USE_EVT || __DOXYGEN__ */
+    lwpkt_evt_fn evt_fn; /*!< Global event function for read and write operation */
+#endif                   /* LWPKT_CFG_USE_EVT || __DOXYGEN__ */
+    uint8_t flags;       /*!< List of flags */
 
     struct {
-        lwpkt_state_t state;                    /*!< Actual packet state machine */
+        lwpkt_state_t state; /*!< Actual packet state machine */
 #if LWPKT_CFG_USE_CRC || __DOXYGEN__
-        lwpkt_crc_t crc;                        /*!< Packet CRC byte */
-#endif /* LWPKT_CFG_USE_CRC || __DOXYGEN__ */
+        lwpkt_crc_t crc; /*!< Packet CRC byte */
+#endif                   /* LWPKT_CFG_USE_CRC || __DOXYGEN__ */
 #if LWPKT_CFG_USE_ADDR || __DOXYGEN__
-        lwpkt_addr_t from;                      /*!< Device address packet is coming from */
-        lwpkt_addr_t to;                        /*!< Device address packet is intended for */
-#endif /* LWPKT_CFG_USE_ADDR || __DOXYGEN__ */
+        lwpkt_addr_t from; /*!< Device address packet is coming from */
+        lwpkt_addr_t to;   /*!< Device address packet is intended for */
+#endif                     /* LWPKT_CFG_USE_ADDR || __DOXYGEN__ */
+#if LWPKT_CFG_USE_FLAGS || __DOXYGEN__
+        uint32_t flags; /*!< Custom flags */
+#endif                  /* LWPKT_CFG_USE_FLAGS || __DOXYGEN__ */
 #if LWPKT_CFG_USE_CMD || __DOXYGEN__
-        uint8_t cmd;                            /*!< Command packet */
-#endif /* LWPKT_CFG_USE_CMD || __DOXYGEN__ */
-        size_t len;                             /*!< Number of bytes to receive */
-        size_t index;                           /*!< General index variable for multi-byte parts of packet */
-    } m;                                        /*!< Module that is periodically reset for next packet */
+        uint8_t cmd;  /*!< Command packet */
+#endif                /* LWPKT_CFG_USE_CMD || __DOXYGEN__ */
+        size_t len;   /*!< Number of bytes to receive */
+        size_t index; /*!< General index variable for multi-byte parts of packet */
+    } m;              /*!< Module that is periodically reset for next packet */
 } lwpkt_t;
 
-lwpktr_t    lwpkt_init(lwpkt_t* pkt, LWRB_VOLATILE lwrb_t* tx_rb, LWRB_VOLATILE lwrb_t* rx_rb);
-lwpktr_t    lwpkt_set_addr(lwpkt_t* pkt, lwpkt_addr_t addr);
-lwpktr_t    lwpkt_read(lwpkt_t* pkt);
-lwpktr_t    lwpkt_write(lwpkt_t* pkt,
+lwpktr_t lwpkt_init(lwpkt_t* pkt, lwrb_t* tx_rb, lwrb_t* rx_rb);
+lwpktr_t lwpkt_set_addr(lwpkt_t* pkt, lwpkt_addr_t addr);
+lwpktr_t lwpkt_read(lwpkt_t* pkt);
+lwpktr_t lwpkt_write(lwpkt_t* pkt,
 #if LWPKT_CFG_USE_ADDR || __DOXYGEN__
-    lwpkt_addr_t to,
+                     lwpkt_addr_t to,
 #endif /* LWPKT_CFG_USE_ADDR || __DOXYGEN__ */
+#if LWPKT_CFG_USE_FLAGS || __DOXYGEN__
+                     uint32_t flags,
+#endif /* LWPKT_CFG_USE_FLAGS || __DOXYGEN__ */
 #if LWPKT_CFG_USE_CMD || __DOXYGEN__
-    uint8_t cmd,
+                     uint8_t cmd,
 #endif /* LWPKT_CFG_USE_CMD || __DOXYGEN__ */
-    const void* data, size_t len);
-lwpktr_t    lwpkt_reset(lwpkt_t* pkt);
-lwpktr_t    lwpkt_process(lwpkt_t* pkt, uint32_t time);
-lwpktr_t    lwpkt_set_evt_fn(lwpkt_t* pkt, lwpkt_evt_fn evt_fn);
+                     const void* data, size_t len);
+lwpktr_t lwpkt_reset(lwpkt_t* pkt);
+lwpktr_t lwpkt_process(lwpkt_t* pkt, uint32_t time);
+lwpktr_t lwpkt_set_evt_fn(lwpkt_t* pkt, lwpkt_evt_fn evt_fn);
+
+/* Functions available as conditional build */
+void lwpkt_set_crc_enabled(lwpkt_t* pkt, uint8_t enable);
+void lwpkt_set_addr_enabled(lwpkt_t* pkt, uint8_t enable);
+void lwpkt_set_addr_extended_enabled(lwpkt_t* pkt, uint8_t enable);
+void lwpkt_set_cmd_enabled(lwpkt_t* pkt, uint8_t enable);
+void lwpkt_set_flags_enabled(lwpkt_t* pkt, uint8_t enable);
 
 /**
  * \brief           Get address from where packet was sent
  * \param[in]       pkt: LwPKT instance
  * \return          Address
  */
-#define lwpkt_get_from_addr(pkt)            (lwpkt_addr_t)   (((pkt) != NULL) ? ((pkt)->m.from) : 0)
+#define lwpkt_get_from_addr(pkt) (lwpkt_addr_t)(((pkt) != NULL) ? ((pkt)->m.from) : 0)
 
 /**
  * \brief           Get address to where packet was sent
  * \param[in]       pkt: LwPKT instance
  * \return          Address
  */
-#define lwpkt_get_to_addr(pkt)              (lwpkt_addr_t)   (((pkt) != NULL) ? ((pkt)->m.to) : 0)
+#define lwpkt_get_to_addr(pkt)   (lwpkt_addr_t)(((pkt) != NULL) ? ((pkt)->m.to) : 0)
 
 /**
  * \brief           Get length of packet
  * \param[in]       pkt: LwPKT instance
  * \return          Number of data bytes in packet
  */
-#define lwpkt_get_data_len(pkt)             (size_t)    (((pkt) != NULL) ? ((pkt)->m.len) : 0)
+#define lwpkt_get_data_len(pkt)  (size_t)(((pkt) != NULL) ? ((pkt)->m.len) : 0)
 
 /**
  * \brief           Get pointer to packet data
  * \param[in]       pkt: LwPKT instance
  * \return          Pointer to data
  */
-#define lwpkt_get_data(pkt)                 (void *)    (((pkt) != NULL) ? ((pkt)->data) : NULL)
+#define lwpkt_get_data(pkt)      (void*)(((pkt) != NULL) ? ((pkt)->data) : NULL)
 
 /**
  * \brief           Get packet command data field
  * \param[in]       pkt: LwPKT instance
  * \return          Command data field
  */
-#define lwpkt_get_cmd(pkt)                  (uint8_t)   (((pkt) != NULL) ? ((pkt)->m.cmd) : 0)
+#define lwpkt_get_cmd(pkt)       (uint8_t)(((pkt) != NULL) ? ((pkt)->m.cmd) : 0)
+
+/**
+ * \brief           Get packet flags
+ * \param[in]       pkt: LwPKT instance
+ * \return          Last received packet flags
+ */
+#define lwpkt_get_flags(pkt)     (uint32_t)(((pkt) != NULL) ? ((pkt)->m.flags) : 0)
 
 /**
  * \brief           Check if packet `to` field address matches device address
  * \param[in]       pkt: LwPKT instance
  * \return          `1` on success, `0` otherwise
  */
-#define lwpkt_is_for_me(pkt)                (((pkt) != NULL) ? ((pkt)->m.to == (pkt)->addr) : 0)
+#define lwpkt_is_for_me(pkt)     (((pkt) != NULL) ? ((pkt)->m.to == (pkt)->addr) : 0)
 
 /**
  * \brief           Check if packet was sent to all devices on network
  * \param[in]       pkt: LwPKT instance
  * \return          `1` if broadcast, `0` otherwise
  */
-#define lwpkt_is_broadcast(pkt)             (((pkt) != NULL) ? ((pkt)->m.to == LWPKT_CFG_ADDR_BROADCAST) : 0)
+#define lwpkt_is_broadcast(pkt)  (((pkt) != NULL) ? ((pkt)->m.to == LWPKT_CFG_ADDR_BROADCAST) : 0)
 
 /**
  * \}