+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Develop
 
+## v1.3.0
+
+- Split CMakeLists.txt files between library and executable
+- Change license year to 2022
+- Add `.clang-format` draft
+- Change license year to 2023
+- Add memory overflow check
+- Add more events for pre and post write & read operations
+- Add more recent version of LwRB
+- Fix compilation error if CRC mode is disabled
+- Add support for dynamic configuration, to support multiple LwPKT instances in one project
+- Add flags support to allow customer user flags in packet
+
 ## v1.2.0
 
 - Added support for events on packet ready, read or write operation